refactor: POC: Moved collector commands to modules
* Each collector sub-command was moved to its own
  module.
